+++ waspc/cli/src/Wasp/Cli/Command/Show/Subcommand.hs
@@ -10,9 +10,9 @@ import Control.Monad.IO.Class (liftIO)
 import Data.Aeson (ToJSON)
 import Data.Aeson.Encode.Pretty (encodePretty)
 import qualified Data.ByteString.Lazy.Char8 as BSL8
-import Wasp.AppSpec.Core.Inspectable (Inspectable, inspect)
 import Wasp.Cli.Command (Command)
 import Wasp.Cli.Command.Show.Table (renderEntriesAsTables)
+import Wasp.Inspectable (Inspectable, inspect)
 
 -- | Describes a `wasp show` subcommand: both its CLI parser and its
 -- implementation are derived from this description.
